# services/gateway/public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'eSapiens Challenge API'
  contact:
    name: 'Guilherme Alves'
    email: guihalves20@gmail.com
  version: '1.0'
paths:
  /users:
    post:
      tags:
        - Users
      summary: 'Cria um usuário no sistema'
      parameters:
        -
          name: name
          in: query
          description: 'Nome do usuário'
          required: true
          schema:
            type: string
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: subscriber
          in: query
          description: 'Identifica se o usuário é assinante (default: false)'
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: OK
        '400':
          description: 'Dados inválidos'
        '403':
          description: 'Autorização negada'
        '500':
          description: 'Erro ao exibir notificação'
...
  '/comments/user/{userID}':
    get:
      tags:
        - Comments
      summary: 'Lista todas as publicações de um usuário'
      operationId: 'App\Http\Controllers\CommentController::indexByUser'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: userID
          in: path
          description: 'ID do user'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /notifications/all:
    get:
      tags:
        - Notifications
      summary: 'Lista todas as notificações do usuário logado'
      operationId: 'App\Http\Controllers\NotificationController::indexAll'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '500':
          description: 'Erro ao exibir notificações'
  /notifications/new:
    get:
      tags:
        - Notifications
      summary: 'Lista todas as novas notificações do usuário logado'
      operationId: 'App\Http\Controllers\NotificationController::indexNew'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '500':
          description: 'Erro ao exibir notificações'
  /notifications:
    get:
      tags:
        - Notifications
      summary: 'Lista todas as notificações do usuário que ainda não expiraram'
      operationId: 'App\Http\Controllers\NotificationController::index'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '500':
          description: 'Erro ao exibir notificações'
  '/notifications/{notificationID}':
    get:
      tags:
        - Notifications
      summary: 'Retorna os dados de uma notificação'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: notificationID
          in: path
          description: 'ID da notificação'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '403':
          description: 'Autorização negada'
        '500':
          description: 'Erro ao exibir notificação'